refactor: 라우팅 경로 수정(onboarding → imageSetup)
