+++ src/pages/imageSetup/types/apis/floorPlan.ts
