--- src/pages/onboarding/components/steps/title/Maintitle.tsx
@@ -1,17 +1,0 @@
-import * as styles from '../StepCommon.css';
-
-interface MainTitleProps {
-  title: string;
-  body?: string;
-}
-
-const MainTitle = ({ title, body }: MainTitleProps) => {
-  return (
-    <div className={styles.mainTextBox}>
-      <span className={styles.title}>{title}</span>
-      {body && <span className={styles.body}>{body}</span>}
-    </div>
-  );
-};
-
-export default MainTitle;
