--- src/pages/onboarding/constants/headerImages.ts
@@ -1,11 +1,0 @@
-import step1 from '@assets/icons/funnelStep1.png';
-import step2 from '@assets/icons/funnelStep2.png';
-import step3 from '@assets/icons/funnelStep3.png';
-import step4 from '@assets/icons/funnelStep4.png';
-
-export const FUNNELHEADER_IMAGES = {
-  1: step1,
-  2: step2,
-  3: step3,
-  4: step4,
-} as const;
