--- src/pages/onboarding/constants/step3.constants.ts
@@ -1,1 +1,0 @@
-export const MAX_MOOD_BOARD_SELECTION = 5;
