--- src/pages/onboarding/hooks/logics/useBottomSheetAddress.ts
@@ -1,15 +1,0 @@
-import { useMutation } from '@tanstack/react-query';
-import { postAddress } from '../../apis/BottomSheetAddress';
-import type { BottomSheetAddressRequest } from '../../types/apis/bottomSheetAddress.types';
-
-export const useBottomSheetAddress = () => {
-  return useMutation({
-    mutationFn: (body: BottomSheetAddressRequest) => postAddress(body),
-    onSuccess: () => {
-      console.log('주소 등록 성공!');
-    },
-    onError: (error) => {
-      console.error('주소 등록 실패', error);
-    },
-  });
-};
