--- src/pages/onboarding/types/OpenSheet.ts
@@ -1,2 +1,0 @@
-export const OPEN_SHEET_KEYS = ['noMatch', 'flip'] as const;
-export type OpenSheetKey = (typeof OPEN_SHEET_KEYS)[number] | null;
