--- src/pages/onboarding/types/apis/bottomSheetAddress.types.ts
@@ -1,4 +1,0 @@
-export type BottomSheetAddressRequest = {
-  sigungu: string;
-  roadName: string;
-};
