--- src/pages/onboarding/types/apis/step2Api.types.ts
@@ -1,14 +1,0 @@
-export interface FloorPlanData {
-  id: number;
-  form: string;
-  structure: string;
-  floorPlanImage: string;
-}
-
-export interface FloorPlanResponse {
-  code: number;
-  msg: string;
-  data: {
-    floorPlanList: FloorPlanData[];
-  };
-}
