--- src/pages/onboarding/types/funnel/index.ts
@@ -1,3 +1,0 @@
-export * from './options';
-export * from './steps';
-export * from './validation';
